Here is a simple benchmark to time relative costs of "headless" vs "plotting" Sunny usage:

```
@time begin
using Sunny
using LinearAlgebra
lat_vecs = [1 1 0; 1 0 1; 0 1 1]' / 2
positions = [[1, 1, 1], [-1, -1, -1]] / 8
cryst = Crystal(lat_vecs, positions)
int = [exchange(1.0*I, Bond(1, 1, [1, 0, 0]))]
end

@time begin
using GLMakie # This line becomes required in the `requires` branch
plot_bonds(cryst, int, (2,2,2); markersize=200)
end
```

On `main` branch, the times are 13s and 42s. On `requires` branch, the times are 6.5s and 50s. So avoiding the `using GLMakie` shaves 50% off the time to do basic Sunny stuff. But then plotting becomes slower, and (if the user eventually wants to make a plot) there is a net cost increase of `56.5 - 55 = 1.5` seconds. I would personally consider this tradeoff worth it, but it should be discussed.