@kbarros 2021-11-09
Cole, incidentally, could you check that the plot the shows up is correct? It's a primitive fcc unit cell, but the dimensions look kind of strange to my eye.