@ColeMiles 2021-11-09
I can confirm similar timings on my machine (16s -> 8s, 54s -> 62s) -- this looks great! I think the hit to plotting time is worth the much faster time for everything else. Plus, it seems like the total time of loading a crystal + plotting it is roughly the same.

I'm wondering what the remaining big-hitters are for loading time. I guess `CrystalInfoFramework` or `Spglib`? We wouldn't want to hide either of these behind `Requires` though.

As for the plot -- it seems to be correct. I think there's just a lot of things going on which make it difficult to see -- the two-atom basis, the small number of unit cells, and I think it's just very difficult to get our brain to ignore the lines which are drawing the primitive unit cell.

Making `positions = [[0, 0, 0]]` and plotting 4x4x4 cells, I can just barely make out the FCC cube:

![image](https://user-images.githubusercontent.com/20207571/140863087-60d0e1d5-f1d9-499c-8ae2-3f56f006d56e.png)

Even there the proportions look off, but if you rotate it around you can see it's a cube.

Anyway, seems good to merge to me.